Remove unnecessary require statements
Since all tasks in fetch.rake depend on the :environment task that means
the entire application environment will be loaded and available for each
task. In other words, all gems from the Gemfile will be already loaded
by the code on line 7 of application.rb.

Leaving the require statements creates problems such as attempting to
require sqlite3 in production, whereas it is only used and available on
development.